Remove KtCompiler from detekt-parser
