+++ detekt-parser/build.gradle.kts
@@ -5,6 +5,5 @@ plugins {
 dependencies {
     api(libs.kotlin.compiler)
     implementation(projects.detektKotlinAnalysisApiStandalone)
-    testImplementation(projects.detektTestUtils)
     testImplementation(libs.assertj.core)
 }
