--- detekt-parser/src/main/kotlin/dev/detekt/parser/KtCompiler.kt
@@ -1,38 +1,0 @@
-package dev.detekt.parser
-
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.vfs.VirtualFileManager
-import com.intellij.pom.PomModel
-import com.intellij.pom.tree.TreeAspect
-import com.intellij.psi.util.PsiUtilCore
-import org.jetbrains.kotlin.analysis.api.standalone.buildStandaloneAnalysisAPISession
-import org.jetbrains.kotlin.platform.jvm.JvmPlatforms
-import org.jetbrains.kotlin.psi.KtFile
-import java.nio.file.Path
-import kotlin.io.path.isRegularFile
-
-open class KtCompiler(
-    protected val project: Project = createDefaultAnalysisAPISession().project,
-) {
-
-    fun compile(path: Path): KtFile {
-        require(path.isRegularFile()) { "Given path '$path' should be a regular file!" }
-
-        val virtualFile = requireNotNull(VirtualFileManager.getInstance().findFileByNioPath(path)) {
-            "$path cannot be retrieved as a VirtualFile"
-        }
-
-        return requireNotNull(PsiUtilCore.getPsiFile(project, virtualFile) as? KtFile) {
-            "$path is not a Kotlin file"
-        }
-    }
-}
-
-private fun createDefaultAnalysisAPISession() = buildStandaloneAnalysisAPISession {
-    registerProjectService(TreeAspect::class.java)
-    registerProjectService(PomModel::class.java, DetektPomModel(project))
-
-    buildKtModuleProvider {
-        platform = JvmPlatforms.defaultJvmPlatform
-    }
-}
