+++ detekt-test-utils/src/test/resources/cases/.gitattributes
