+++ detekt-test-utils/src/test/resources/cases/DefaultCrLf.kt
