+++ detekt-test-utils/src/test/resources/cases/DefaultLf.kt
