+++ detekt-test-utils/src/test/resources/css/test.css
